requirements.txt
aniso8601==9.0.1
asgiref==3.7.2
certifi==2023.5.7
charset-normalizer==3.1.0
coreapi==2.3.3
coreschema==0.0.4
Django==4.2.3
django-autoslug==1.9.9
django-filter==23.2
django-graphql-jwt==0.3.4
django-jazzmin==2.6.0
django-widget-tweaks==1.4.12
djangorestframework==3.14.0
drf-yasg==1.21.6
graphene==3.2.2
graphene-django==3.1.2
graphql-core==3.2.3
graphql-relay==3.2.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.3
packaging==23.1
Pillow==10.0.0
promise==2.3
psycopg2-binary==2.9.6
PyJWT==2.7.0
pytz==2023.3
PyYAML==6.0
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
text-unidecode==1.3
typing_extensions==4.7.1
uritemplate==4.1.1
urllib3==2.0.3
